@@ -1,262 +1,129 @@
-#include "DataDependencyGraph.hpp"
-
-using namespace llvm;
+#include "DataDependencyGraph.hh"
 
 char pdg::DataDependencyGraph::ID = 0;
 
-void pdg::DataDependencyGraph::initializeMemoryDependencyPasses()
-{
-  steenAA = &getAnalysis<CFLSteensAAWrapperPass>().getResult();
-  andersAA = &getAnalysis<CFLAndersAAWrapperPass>().getResult();
-  MD = &getAnalysis<MemoryDependenceWrapperPass>().getMemDep();
-}
+using namespace llvm;
 
-void pdg::DataDependencyGraph::constructFuncMapAndCreateFunctionEntry()
+bool pdg::DataDependencyGraph::runOnModule(Module &M)
 {
-  auto& pdgUtils = PDGUtils::getInstance();
-  if (pdgUtils.getFuncMap()[Func]->getEntryW() == nullptr)
+  ProgramGraph &g = ProgramGraph::getInstance();
+  if (!g.isBuild())
   {
-    InstructionWrapper *entryW = new InstructionWrapper(Func, GraphNodeType::ENTRY);
-    pdgUtils.getFuncInstWMap()[Func].insert(entryW);
-    pdgUtils.getFuncMap()[Func]->setEntryW(entryW);
+    g.build(M);
+    // TODO: add comment
+    g.bindDITypeToNodes(M);
   }
-}
-
-void pdg::DataDependencyGraph::collectDataDependencyInFunc()
-{
-  for (inst_iterator instIt = inst_begin(Func); instIt != inst_end(Func); ++instIt)
+  
+  for (auto &F : M)
   {
-    getNodeByData(&*instIt);
-    Instruction *inst = dyn_cast<Instruction>(&*instIt);
-    collectDefUseDependency(inst);
-    collectCallInstDependency(inst);
-
-    if (isa<LoadInst>(inst))
+    if (F.isDeclaration() || F.empty())
+      continue;
+    _mem_dep_res = &getAnalysis<MemoryDependenceWrapperPass>(F).getMemDep();
+    // setup alias query interface for each function
+    for (auto inst_iter = inst_begin(F); inst_iter != inst_end(F); inst_iter++)
     {
-      collectReadFromDependency(inst);
-      collectRAWDependency(inst);
-      collectNonLocalDependency(inst);
+      addDefUseEdges(*inst_iter);
+      addRAWEdges(*inst_iter);
+      addAliasEdges(*inst_iter);
     }
   }
-}
-
-void pdg::DataDependencyGraph::collectAliasDependencies()
-{
-  auto &pdgUtils = PDGUtils::getInstance();
-  auto funcMap = pdgUtils.getFuncMap()[Func];
-  auto instMap = pdgUtils.getInstMap();
-  auto storeVec = funcMap->getStoreInstList();
-  auto loadVec = funcMap->getLoadInstList();
-  auto castVec = funcMap->getCastInstList();
-
-  for (StoreInst *si : storeVec)
-  {
-    MemoryLocation s_loc = MemoryLocation::get(si);
-    for (LoadInst *li : loadVec)
-    {
-      MemoryLocation l_loc = MemoryLocation::get(li);
-      AliasResult andersAAResult = andersAA->query(s_loc, l_loc);
-      // AliasResult steensAAResult = steenAA->alias(s_loc, l_loc);
-      if (andersAAResult != NoAlias)
-      {
-        InstructionWrapper *loadInstW = instMap[li];
-        InstructionWrapper *storeInstW = instMap[si];
-        DDG->addDependency(storeInstW, loadInstW, DependencyType::DATA_ALIAS);
-        DDG->addDependency(loadInstW, storeInstW, DependencyType::DATA_ALIAS);
-      }
-    }
-    
-    for (StoreInst *si1 : storeVec) {
-      if (si == si1)
-        continue;
-      MemoryLocation s1_loc = MemoryLocation::get(si1);
-      AliasResult andersAAResult = andersAA->query(s_loc, s1_loc);
-      if (andersAAResult != NoAlias) {
-        InstructionWrapper *store1InstW = PDGUtils::getInstance().getInstMap()[si];
-        InstructionWrapper *store2InstW = PDGUtils::getInstance().getInstMap()[si1];
-        DDG->addDependency(store1InstW, store2InstW, DependencyType::DATA_ALIAS);
-        DDG->addDependency(store2InstW, store1InstW, DependencyType::DATA_ALIAS);
-      }
-    }
-  }
-
-  // add alias info for two load instructions
-  for (LoadInst *li1 : loadVec)
-  {
-    for (LoadInst *li2 : loadVec)
-    {
-      if (li1 == li2)
-      {
-        continue;
-      }
-      MemoryLocation li1_loc = MemoryLocation::get(li1);
-      MemoryLocation li2_loc = MemoryLocation::get(li2);
-
-      Type *li1LocTy = li1->getPointerOperandType();
-      Type *li2LocTy = li2->getPointerOperandType();
-
-      if (li1 != li2)
-      {
-        continue;
-      }
-      AliasResult AA_result = andersAA->query(li1_loc, li2_loc);
-      if (AA_result != NoAlias)
-      {
-        InstructionWrapper *loadInstW1 = instMap[li1];
-        InstructionWrapper *loadInstW2 = instMap[li2];
-        DDG->addDependency(loadInstW1, loadInstW2, DependencyType::DATA_ALIAS);
-      }
-    }
-  }
-
-  for (CastInst *csi : castVec)
-  {
-    errs() << *csi << "\n";
-    auto srcInst = dyn_cast<Instruction>(csi->getOperand(0));
-    auto destInst = csi;
-
-    InstructionWrapper *srcInstW = instMap[srcInst];
-    InstructionWrapper *destInstW = instMap[destInst];
-    DDG->addDependency(srcInstW, destInstW, DependencyType::DATA_ALIAS);
-    DDG->addDependency(destInstW, srcInstW, DependencyType::DATA_ALIAS);
-  }
-}
-
-void pdg::DataDependencyGraph::collectReadFromDependency(llvm::Instruction *inst)
-{
-  if (LoadInst *li = dyn_cast<LoadInst>(inst))
-  {
-    if (Instruction *pInst = dyn_cast<Instruction>(li->getPointerOperand()))
-    {
-      DDG->addDependency(PDGUtils::getInstance().getInstMap()[pInst],
-                         PDGUtils::getInstance().getInstMap()[inst],
-                         DependencyType::DATA_READ);
-    }
-  }
-}
-
-void pdg::DataDependencyGraph::collectDefUseDependency(llvm::Instruction *inst)
-{
-  // check for def-use dependencies
-  for (Instruction::const_op_iterator cuit = inst->op_begin();
-       cuit != inst->op_end(); ++cuit)
-  {
-    if (Instruction *pInst = dyn_cast<Instruction>(*cuit))
-    {
-      // add info flow from the instruction to current instruction
-      DDG->addDependency(PDGUtils::getInstance().getInstMap()[pInst],
-                         PDGUtils::getInstance().getInstMap()[inst],
-                         DependencyType::DATA_DEF_USE);
-    }
-  }
-}
-
-void pdg::DataDependencyGraph::collectCallInstDependency(llvm::Instruction *inst)
-{
-  if (CallInst *callInst = dyn_cast<CallInst>(inst))
-  {
-    if (callInst->getCalledFunction() != nullptr && callInst->getCalledFunction()->isDeclaration())
-      return;
-
-    for (auto arg_iter = callInst->arg_begin(); arg_iter != callInst->arg_end(); ++arg_iter)
-    {
-      if (Instruction *tmpInst = dyn_cast<Instruction>(&*arg_iter))
-      {
-        // DDG->addDependency(PDGUtils::getInstance().getInstMap()[tmpInst],
-        //                    PDGUtils::getInstance().getInstMap()[inst],
-        //                    DependencyType::DATA_CALL_PARA);
-        DDG->addDependency(PDGUtils::getInstance().getInstMap()[inst],
-                           PDGUtils::getInstance().getInstMap()[tmpInst],
-                           DependencyType::DATA_CALL_PARA);
-      }
-    }
-  }
-}
-
-std::vector<Instruction *> pdg::DataDependencyGraph::getRAWDepList(Instruction *pLoadInst)
-{
-  std::vector<Instruction *> _flowdep_set;
-  std::vector<StoreInst *> StoreVec = PDGUtils::getInstance().getFuncMap()[Func]->getStoreInstList();
-  // for each Load Instruction, find related Store Instructions(alias considered)
-  LoadInst *LI = dyn_cast<LoadInst>(pLoadInst);
-  MemoryLocation LI_Loc = MemoryLocation::get(LI);
-  for (StoreInst *SI : StoreVec)
-  {
-    MemoryLocation SI_Loc = MemoryLocation::get(SI);
-    AliasResult andersAAResult = andersAA->query(LI_Loc, SI_Loc);
-    AliasResult steensAAResult = steenAA->query(LI_Loc, SI_Loc);
-    if (andersAAResult != NoAlias || steensAAResult != NoAlias)
-    {
-      _flowdep_set.push_back(SI);
-    }
-  }
-  return _flowdep_set;
-}
-
-void pdg::DataDependencyGraph::collectRAWDependency(llvm::Instruction *inst)
-{
-  // dealing with dependencies in a function
-  std::vector<Instruction *> flowdep_set = getRAWDepList(inst);
-
-  for (unsigned i = 0; i < flowdep_set.size(); i++)
-  {
-    DDG->addDependency(PDGUtils::getInstance().getInstMap()[flowdep_set[i]],
-                       PDGUtils::getInstance().getInstMap()[inst],
-                       DependencyType::DATA_RAW);
-  }
-
-  flowdep_set.clear();
-}
-
-void pdg::DataDependencyGraph::collectNonLocalDependency(llvm::Instruction *inst)
-{
-  // dealing with non local pointer dependency, nonLocalPointer dep is stored in result small vector
-  SmallVector<NonLocalDepResult, 20> result;
-  // the return result is NonLocalDepResult. can use getAddress function
-  MD->getNonLocalPointerDependency(inst, result);
-  // now result stores all possible
-  for (NonLocalDepResult &I : result)
-  {
-    const MemDepResult &nonLocal_res = I.getResult();
-    InstructionWrapper *itInst = PDGUtils::getInstance().getInstMap()[inst];
-    InstructionWrapper *parentInst = PDGUtils::getInstance().getInstMap()[nonLocal_res.getInst()];
-
-    if (nonLocal_res.getInst() != nullptr)
-    {
-      DDG->addDependency(itInst, parentInst, DependencyType::DATA_GENERAL);
-    }
-    // ignore nullptr return res
-  }
-}
-
-pdg::DependencyNode<pdg::InstructionWrapper> *pdg::DataDependencyGraph::getNodeByData(Instruction *inst)
-{
-  return DDG->getNodeByData(PDGUtils::getInstance().getInstMap()[inst]);
-}
-
-typename pdg::DependencyNode<pdg::InstructionWrapper>::DependencyLinkList pdg::DataDependencyGraph::getNodeDepList(Instruction *inst)
-{
-  return DDG->getNodeDepList(PDGUtils::getInstance().getInstMap()[inst]);
-}
-
-bool pdg::DataDependencyGraph::runOnFunction(Function &F)
-{
-  Func = &F;
-  PDGUtils::getInstance().constructFuncMap(*F.getParent());
-  initializeMemoryDependencyPasses();
-  constructFuncMapAndCreateFunctionEntry();
-  collectDataDependencyInFunc();
-  collectAliasDependencies();
   return false;
 }
 
-void pdg::DataDependencyGraph::getAnalysisUsage(AnalysisUsage &AU) const
+
+void pdg::DataDependencyGraph::addAliasEdges(Instruction &inst)
 {
-  AU.addRequired<MemoryDependenceWrapperPass>();
-  AU.addRequired<CFLSteensAAWrapperPass>();
-  AU.addRequired<CFLAndersAAWrapperPass>();
-  AU.setPreservesAll();
+  ProgramGraph &g = ProgramGraph::getInstance();
+  Function* func = inst.getFunction();
+  for (auto inst_iter = inst_begin(func); inst_iter != inst_end(func); inst_iter++)
+  {
+    if (&inst == &*inst_iter)
+      continue;
+    
+    auto alias_result = queryAliasUnderApproximate(inst, *inst_iter);
+    if (alias_result != NoAlias)
+    {
+      Node* src = g.getNode(inst);
+      Node* dst = g.getNode(*inst_iter);
+      if (src == nullptr || dst == nullptr)
+        continue;
+      src->addNeighbor(*dst, EdgeType::DATA_ALIAS);
+    }
+  }
 }
 
-static RegisterPass<pdg::DataDependencyGraph>
-    DDG("ddg", "Data Dependency Graph Construction", false, true);
+void pdg::DataDependencyGraph::addDefUseEdges(Instruction &inst)
+{
+  ProgramGraph &g = ProgramGraph::getInstance();
+  for (auto user : inst.users())
+  {
+    Node *src = g.getNode(inst);
+    Node *dst = g.getNode(*user);
+    if (src == nullptr || dst == nullptr)
+      continue;
+    EdgeType edge_type = EdgeType::DATA_DEF_USE;
+    if (dst->getNodeType() == GraphNodeType::ANNO_VAR)
+      edge_type = EdgeType::ANNO_VAR;
+    if (dst->getNodeType() == GraphNodeType::ANNO_GLOBAL)
+      edge_type = EdgeType::ANNO_GLOBAL;
+    src->addNeighbor(*dst, edge_type);
+  }
+}
+
+void pdg::DataDependencyGraph::addRAWEdges(Instruction &inst)
+{
+  if (!isa<LoadInst>(&inst))
+    return;
+
+  ProgramGraph &g = ProgramGraph::getInstance();
+  auto dep_res = _mem_dep_res->getDependency(&inst);
+  auto dep_inst = dep_res.getInst();
+
+  if (!dep_inst)
+    return;
+  if (!isa<StoreInst>(dep_inst))
+    return;
+
+  Node *src = g.getNode(inst);
+  Node *dst = g.getNode(*dep_inst);
+  if (src == nullptr || dst == nullptr)
+    return;
+  dst->addNeighbor(*src, EdgeType::DATA_RAW);
+}
+
+AliasResult pdg::DataDependencyGraph::queryAliasUnderApproximate(Value &v1, Value &v2)
+{
+  if (!v1.getType()->isPointerTy() || !v2.getType()->isPointerTy())
+    return NoAlias;
+  // check bit cast
+  if (BitCastInst *bci = dyn_cast<BitCastInst>(&v1))
+  {
+    if (bci->getOperand(0) == &v2)
+      return MustAlias;
+  }
+  // handle load instruction
+  if (LoadInst *li = dyn_cast<LoadInst>(&v1))
+  {
+    auto load_addr = li->getPointerOperand();
+    for (auto user : load_addr->users())
+    {
+      if (StoreInst *si = dyn_cast<StoreInst>(user))
+      {
+        if (si->getPointerOperand() == load_addr)
+        {
+          if (si->getValueOperand() == &v2)
+            return MustAlias;
+        }
+      }
+    }
+  }
+}
+
+  void pdg::DataDependencyGraph::getAnalysisUsage(AnalysisUsage & AU) const
+  {
+    AU.addRequired<MemoryDependenceWrapperPass>();
+    AU.setPreservesAll();
+  }
+
+  static RegisterPass<pdg::DataDependencyGraph>
+      DDG("ddg", "Data Dependency Graph Construction", false, true);
